--- a/.gitignore
+++ b/.gitignore
@@ -15,3 +15,5 @@
 .env
 npm-debug.log
 LW_frontend.code-workspace
+
+.vercel