--- a/constants.js
+++ b/constants.js
@@ -22,9 +22,9 @@ exports.PUBLIC_BASE_URL = PUBLIC_BASE_URL
 const location = typeof window !== 'undefined' && window.location
 exports.GRAPHQL_URI = location
   ? `${location.protocol}//${location.hostname}${location.port ? `:${location.port}` : ''}/graphql`
-  : PUBLIC_BASE_URL
-    ? `${PUBLIC_BASE_URL}/graphql`
-    : `http://127.0.0.1:${process.env.PORT || 3000}/graphql`
+  : NEXT_PUBLIC_VERCEL_URL
+    ? `${PUBLIC_BASE_URL}/graphql` // on vercel there is no local http server, we need to go via public url
+    : `http://127.0.0.1:${process.env.PORT || 3000}/graphql` // for fast access on heroku and for localhost
 
 exports.GA_TRACKING_ID = NEXT_PUBLIC_GA_TRACKING_ID