--- a/lib/apolloClient.js
+++ b/lib/apolloClient.js
@@ -4,6 +4,7 @@ import {getDataFromTree} from '@apollo/client/react/ssr'
 import deepMerge from './deepMerge'
 
 import {GRAPHQL_URI} from '../constants'
+import baseQueries from './baseQueries'
 
 export const APOLLO_STATE_PROP_NAME = '__APOLLO_STATE__'
 
@@ -85,6 +86,36 @@ export function withInitialProps(Page) {
     }
   }
   return Page
+}
+
+export function createGetStaticProps({ pageQuery, getVariables } = {}) {
+  return async function getStaticProps({ params }) {
+    const apolloClient = initializeApollo()
+
+    const { locale } = params
+    await Promise.all(
+      baseQueries
+        .map(query => ({
+          query,
+          variables: {
+            locale
+          }
+        }))
+        .concat(pageQuery ? {
+          query: pageQuery,
+          variables: {
+            locale,
+            ...getVariables ? getVariables({ params }) : undefined
+          }
+        } : [])
+        .map(options => apolloClient.query(options))
+    )
+
+    return addApolloState(apolloClient, {
+      props: {},
+      revalidate: 60,
+    })
+  }
 }
 
 export function useApollo(pageProps) {