--- a/lib/apolloClient.js
+++ b/lib/apolloClient.js
@@ -88,12 +88,12 @@ export function withInitialProps(Page) {
   return Page
 }
 
-export function createGetStaticProps({ pageQuery, getVariables } = {}) {
+export function createGetStaticProps({ pageQuery, getVariables, isNotFound } = {}) {
   return async function getStaticProps({ params }) {
     const apolloClient = initializeApollo()
 
     const { locale } = params
-    await Promise.all(
+    const datas = await Promise.all(
       baseQueries
         .map(query => ({
           query,
@@ -112,6 +112,9 @@ export function createGetStaticProps({ pageQuery, getVariables } = {}) {
     )
 
     return addApolloState(apolloClient, {
+      notFound: isNotFound
+        ? isNotFound(datas[datas.length - 1])
+        : false,
       props: {},
       revalidate: 60,
     })