--- a/lib/apolloClient.js
+++ b/lib/apolloClient.js
@@ -92,7 +92,9 @@ export function createGetStaticProps({ pageQuery, getVariables, isNotFound } = {
   return async function getStaticProps({ params }) {
     const apolloClient = initializeApollo()
 
-    const { locale } = params
+    // 404 page has no params
+    const locale = params?.locale || 'de'
+
     const datas = await Promise.all(
       baseQueries
         .map(query => ({