--- a/lib/apolloClient.js
+++ b/lib/apolloClient.js
@@ -5,6 +5,7 @@ import deepMerge from './deepMerge'
 
 import {GRAPHQL_URI} from '../constants'
 import baseQueries from './baseQueries'
+import {locales} from '../constants'
 
 export const APOLLO_STATE_PROP_NAME = '__APOLLO_STATE__'
 
@@ -88,13 +89,18 @@ export function withInitialProps(Page) {
   return Page
 }
 
-export function createGetStaticProps({ pageQuery, getVariables, isNotFound } = {}) {
+export function createGetStaticProps({ pageQuery, getVariables, isNotFound, defaultLocale } = {}) {
   return async function getStaticProps({ params }) {
+    const locale = params?.locale || defaultLocale
+    if (!locales.includes(locale)) {
+      return {
+        notFound: true,
+        props: {},
+        revalidate: 60
+      }
+    }
+
     const apolloClient = initializeApollo()
-
-    // 404 page has no params
-    const locale = params?.locale || 'de'
-
     const datas = await Promise.all(
       baseQueries
         .map(query => ({