--- a/lib/apolloClient.js
+++ b/lib/apolloClient.js
@@ -4,8 +4,6 @@ import {getDataFromTree} from '@apollo/client/react/ssr'
 import deepMerge from './deepMerge'
 
 import {GRAPHQL_URI} from '../constants'
-import baseQueries from './baseQueries'
-import {locales} from '../constants'
 
 export const APOLLO_STATE_PROP_NAME = '__APOLLO_STATE__'
 
@@ -23,6 +21,16 @@ const dataIdFromObject = object => {
 
 let apolloClient
 
+export function createCache() {
+  return new InMemoryCache({
+    possibleTypes: {
+      Person: ['Parliamentarian','Guest'],
+      Entity: ['Parliamentarian','Organisation','Guest','LobbyGroup','Branch']
+    },
+    dataIdFromObject
+  })
+}
+
 function createApolloClient() {
   return new ApolloClient({
     ssrMode: typeof window === 'undefined',
@@ -30,13 +38,7 @@ function createApolloClient() {
       uri: GRAPHQL_URI,
       credentials: 'same-origin',
     }),
-    cache: new InMemoryCache({
-      possibleTypes: {
-        Person: ['Parliamentarian','Guest'],
-        Entity: ['Parliamentarian','Organisation','Guest','LobbyGroup','Branch']
-      },
-      dataIdFromObject
-    }),
+    cache: createCache(),
   })
 }
 
@@ -87,46 +89,6 @@ export function withInitialProps(Page) {
     }
   }
   return Page
-}
-
-export function createGetStaticProps({ pageQuery, getVariables, isNotFound, defaultLocale } = {}) {
-  return async function getStaticProps({ params }) {
-    const locale = params?.locale || defaultLocale
-    if (!locales.includes(locale)) {
-      return {
-        notFound: true,
-        props: {},
-        revalidate: 60
-      }
-    }
-
-    const apolloClient = initializeApollo()
-    const datas = await Promise.all(
-      baseQueries
-        .map(query => ({
-          query,
-          variables: {
-            locale
-          }
-        }))
-        .concat(pageQuery ? {
-          query: pageQuery,
-          variables: {
-            locale,
-            ...getVariables ? getVariables({ params }) : undefined
-          }
-        } : [])
-        .map(options => apolloClient.query(options))
-    )
-
-    return addApolloState(apolloClient, {
-      notFound: isNotFound
-        ? isNotFound(datas[datas.length - 1])
-        : false,
-      props: {},
-      revalidate: 60,
-    })
-  }
 }
 
 export function useApollo(pageProps) {