--- a/lib/apolloClientSchemaLink.js
+++ b/lib/apolloClientSchemaLink.js
@@ -6,14 +6,12 @@ import {makeExecutableSchema} from 'graphql-tools'
 import {createCache, addApolloState} from './apolloClient'
 import baseQueries from './baseQueries'
 
-export function createGetStaticProps({ pageQuery, getVariables, isNotFound, defaultLocale } = {}) {
+export function createGetStaticProps({ pageQuery, getVariables, getCustomStaticProps, defaultLocale } = {}) {
   return async function getStaticProps({ params }) {
     const locale = params?.locale || defaultLocale
     if (!locales.includes(locale)) {
       return {
-        notFound: true,
-        props: {},
-        revalidate: 60
+        notFound: true
       }
     }
 
@@ -49,11 +47,11 @@ export function createGetStaticProps({ pageQuery, getVariables, isNotFound, defa
     )
 
     return addApolloState(apolloClient, {
-      notFound: isNotFound
-        ? isNotFound(datas[datas.length - 1])
-        : false,
       props: {},
       revalidate: 60,
+      ...getCustomStaticProps
+        ? getCustomStaticProps(datas[datas.length - 1], { params })
+        : undefined
     })
   }
 }