--- a/lib/middleware.js
+++ b/lib/middleware.js
@@ -1,9 +1,0 @@
-import {applyMiddleware, compose} from 'redux'
-
-export default function createMiddleware (clientMiddleware) {
-  const middleware = applyMiddleware(clientMiddleware)
-  if (process.browser && window.devToolsExtension) {
-    return compose(middleware, window.devToolsExtension())
-  }
-  return middleware
-}