--- a/lib/reducer.js
+++ b/lib/reducer.js
@@ -1,7 +1,0 @@
-import {combineReducers} from 'redux'
-
-export default function getReducer (client) {
-  return combineReducers({
-    apollo: client.reducer()
-  })
-}