--- a/next.config.js
+++ b/next.config.js
@@ -1,3 +1,7 @@
+const {
+  localeSegment
+} = require('./constants')
+
 module.exports = {
   future: {
     webpack5: true,
@@ -14,6 +18,37 @@ module.exports = {
     }
 
     return config
+  },
+  async rewrites() {
+    return [
+      {
+        source: '/',
+        destination: '/api/language',
+      },
+      {
+        source: '/graphql',
+        destination: '/api/graphql',
+      },
+    ]
+  },
+  async redirects() {
+    return [
+      {
+        source: `/${localeSegment}/search/daten`,
+        destination: '/:locale/search',
+        permanent: true,
+      },
+      {
+        source: `/${localeSegment}/search/daten/:term`,
+        destination: '/:locale/search?term=:term',
+        permanent: true,
+      },
+      {
+        source: '/graphiql',
+        destination: '/graphql',
+        permanent: false
+      }
+    ]
   },
   poweredByHeader: false,
   useFileSystemPublicRoutes: true