--- a/pages/404.js
+++ b/pages/404.js
@@ -3,6 +3,7 @@ import React from 'react'
 import Frame, {Center} from 'src/components/Frame'
 import MetaTags from 'src/components/MetaTags'
 import {H1, P} from 'src/components/Styled'
+import {createGetStaticProps} from 'lib/apolloClient'
 
 export const NotFound = () => <>
   <MetaTags locale='de' title='404' />
@@ -14,13 +15,18 @@ export const NotFound = () => <>
   </Center>
 </>
 
-const Page = ({ serverContext }) => {
-  if (serverContext) {
-    serverContext.res.statusCode = 404
-  }
+const Page = () => {
   return <Frame>
     <NotFound />
   </Frame>
+}
+
+export const getStaticProps = createGetStaticProps()
+export async function getStaticPaths() {
+  return {
+    paths: [],
+    fallback: 'blocking'
+  }
 }
 
 export default Page