--- a/pages/api/language.js
+++ b/pages/api/language.js
@@ -1,0 +1,10 @@
+const acceptLanguage = require('accept-language')
+const {
+  locales
+} = require('../../constants')
+
+acceptLanguage.languages(locales)
+
+export default function handler(req, res) {
+  res.redirect(302, `/${acceptLanguage.get(req.headers['accept-language'])}`)
+}