… README.md
… api.js
… graphql/index.js
… lib/README.md
… lib/initApollo.js
… lib/introspectionQueryResultData.json
… lib/withApolloClient.js
… package.json
… pages/[locale]/daten/lobbygruppe/index.js
… pages/[locale]/daten/organisation/[id]/[[...name]].js
… pages/[locale]/daten/parlamentarier/[id]/[[...name]].js
… pages/[locale]/daten/parlamentarier/index.js
… pages/[locale]/daten/zutrittsberechtigter/[id]/[[...name]].js
… pages/[locale]/daten/zutrittsberechtigter/index.js
… pages/[locale]/index.js
… pages/[locale]/search.js
… pages/_app.js
… pages/_document.js
… pages/api/graphql.js
… public/google8a64872b44ccae7c.html
… server.js
… src/components/BlockRegion.js
… src/components/Frame/Footer.js
… src/components/Frame/Header.js
… src/components/Frame/LoadingBar.js
… src/components/Frame/index.js
… src/components/ListView.js
… src/components/Message.js
… src/components/MetaTags.js
… src/components/Styled.js