This code change gets ride of the express server and allows deploys to vercel.com. Currently running under [lobbywatch.vercel.app](https://lobbywatch.vercel.app/) for testing. The new code can still also run on Heroku, currently also deployed on [lobbywatch-staging.herokuapp.com](https://lobbywatch-staging.herokuapp.com/).

Next step is to refactor the data fetching to enable ISR—incremental static regeneration (uses revalidate in getStaticProps). This should provide a signification performance boost on repeat requests.

## Performance Before ISR Optimisation 

<details>
 <summary>Generally the new and old code run at similar speeds. Varying based on network and backend condition between 700 and 1000ms for a SSR render of a parliamentarian detail page.</summary>

### New Code, Vercel Pro Plan, Paris (cdg1), USD 20/User/Month

```
$ ab -n 100 -c 1 https://lobbywatch.vercel.app/de/daten/parlamentarier/364/Marianne%20Binder-Keller
This is ApacheBench, Version 2.3 <$Revision: 1879490 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking lobbywatch.vercel.app (be patient).....done


Server Software:        Vercel
Server Hostname:        lobbywatch.vercel.app
Server Port:            443
SSL/TLS Protocol:       TLSv1.2,ECDHE-RSA-AES256-GCM-SHA384,2048,256
Server Temp Key:        ECDH X25519 253 bits
TLS Server Name:        lobbywatch.vercel.app

Document Path:          /de/daten/parlamentarier/364/Marianne%20Binder-Keller
Document Length:        234061 bytes

Concurrency Level:      1
Time taken for tests:   69.483 seconds
Complete requests:      100
Failed requests:        0
Total transferred:      23453700 bytes
HTML transferred:       23406100 bytes
Requests per second:    1.44 [#/sec] (mean)
Time per request:       694.832 [ms] (mean)
Time per request:       694.832 [ms] (mean, across all concurrent requests)
Transfer rate:          329.63 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:       88   99  14.7     94     194
Processing:   410  596 219.6    522    1755
Waiting:      365  553 220.0    480    1715
Total:        505  695 220.1    616    1846

Percentage of the requests served within a certain time (ms)
  50%    616
  66%    671
  75%    729
  80%    767
  90%    928
  95%   1202
  98%   1643
  99%   1846
 100%   1846 (longest request)
```

### New Code, Heroku Standard 1X Dyno, Europe (AWS Dublin), USD 25/Month

```
$ ab -n 100 -c 1 https://lobbywatch-staging.herokuapp.com/de/daten/parlamentarier/364/Marianne%20Binder-Keller
This is ApacheBench, Version 2.3 <$Revision: 1879490 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking lobbywatch-staging.herokuapp.com (be patient).....done


Server Software:        Cowboy
Server Hostname:        lobbywatch-staging.herokuapp.com
Server Port:            443
SSL/TLS Protocol:       TLSv1.2,ECDHE-RSA-AES128-GCM-SHA256,2048,128
Server Temp Key:        ECDH P-256 256 bits
TLS Server Name:        lobbywatch-staging.herokuapp.com

Document Path:          /de/daten/parlamentarier/364/Marianne%20Binder-Keller
Document Length:        234036 bytes

Concurrency Level:      1
Time taken for tests:   93.674 seconds
Complete requests:      100
Failed requests:        0
Total transferred:      23427300 bytes
HTML transferred:       23403600 bytes
Requests per second:    1.07 [#/sec] (mean)
Time per request:       936.735 [ms] (mean)
Time per request:       936.735 [ms] (mean, across all concurrent requests)
Transfer rate:          244.23 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:      107  120   7.5    122     135
Processing:   653  816 183.8    775    1806
Waiting:      496  666 185.1    627    1647
Total:        764  937 184.6    895    1931

Percentage of the requests served within a certain time (ms)
  50%    895
  66%    934
  75%    948
  80%    958
  90%   1030
  95%   1044
  98%   1866
  99%   1931
 100%   1931 (longest request)
```

### Old Code, Heroku Standard 1X Dyno, Europe (AWS Dublin), USD 25/Month

```
$ ab -n 100 -c 1 https://lobbywatch.ch/de/daten/parlamentarier/364/Marianne%20Binder-Keller
This is ApacheBench, Version 2.3 <$Revision: 1879490 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking lobbywatch.ch (be patient).....done


Server Software:        Cowboy
Server Hostname:        lobbywatch.ch
Server Port:            443
SSL/TLS Protocol:       TLSv1.2,ECDHE-RSA-AES128-GCM-SHA256,2048,128
Server Temp Key:        ECDH P-256 256 bits
TLS Server Name:        lobbywatch.ch

Document Path:          /de/daten/parlamentarier/364/Marianne%20Binder-Keller
Document Length:        238609 bytes

Concurrency Level:      1
Time taken for tests:   73.075 seconds
Complete requests:      100
Failed requests:        47
   (Connect: 0, Receive: 0, Length: 47, Exceptions: 0)
Total transferred:      23887417 bytes
HTML transferred:       23861417 bytes
Requests per second:    1.37 [#/sec] (mean)
Time per request:       730.752 [ms] (mean)
Time per request:       730.752 [ms] (mean, across all concurrent requests)
Transfer rate:          319.23 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:      114  129   8.4    129     147
Processing:   509  601  48.1    596     812
Waiting:      365  447  45.6    441     669
Total:        630  731  51.5    724     933

Percentage of the requests served within a certain time (ms)
  50%    724
  66%    750
  75%    764
  80%    768
  90%    794
  95%    819
  98%    892
  99%    933
 100%    933 (longest request)
 ```
 </details>