@tpreusse 2021-04-24
## Performance After ISR Optimisation 

Basically 4x faster with Vercel and ISR 🎉 

<details>
 <summary>Now the real performance improvements start happening. On Heroku we're now down to 284ms for a parliamentarian detail page. And on Vercel, with an edge cache in Zurich, down to **177ms**. </summary>

### New Code, Vercel Pro Plan, Paris (cdg1), USD 20/User/Month

```
$ ab -n 100 -l -c 1 https://lobbywatch.vercel.app/de/daten/parlamentarier/364/Marianne%20Binder-Keller
This is ApacheBench, Version 2.3 <$Revision: 1879490 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking lobbywatch.vercel.app (be patient).....done


Server Software:        Vercel
Server Hostname:        lobbywatch.vercel.app
Server Port:            443
SSL/TLS Protocol:       TLSv1.2,ECDHE-RSA-AES256-GCM-SHA384,2048,256
Server Temp Key:        ECDH X25519 253 bits
TLS Server Name:        lobbywatch.vercel.app

Document Path:          /de/daten/parlamentarier/364/Marianne%20Binder-Keller
Document Length:        Variable

Concurrency Level:      1
Time taken for tests:   17.697 seconds
Complete requests:      100
Failed requests:        0
Total transferred:      13496761 bytes
HTML transferred:       13449214 bytes
Requests per second:    5.65 [#/sec] (mean)
Time per request:       176.965 [ms] (mean)
Time per request:       176.965 [ms] (mean, across all concurrent requests)
Transfer rate:          744.80 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:       87  106  38.0     94     303
Processing:    58   71  37.5     62     319
Waiting:       28   40  37.5     31     288
Total:        147  177  54.7    157     439

Percentage of the requests served within a certain time (ms)
  50%    157
  66%    164
  75%    167
  80%    170
  90%    240
  95%    325
  98%    386
  99%    439
 100%    439 (longest request)
```

### New Code, Heroku Standard 1X Dyno, Europe (AWS Dublin), USD 25/Month

```
$ ab -n 100 -l -c 1 https://lobbywatch-staging.herokuapp.com/de/daten/parlamentarier/364/Marianne%20Binder-Keller
This is ApacheBench, Version 2.3 <$Revision: 1879490 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking lobbywatch-staging.herokuapp.com (be patient).....done


Server Software:        Cowboy
Server Hostname:        lobbywatch-staging.herokuapp.com
Server Port:            443
SSL/TLS Protocol:       TLSv1.2,ECDHE-RSA-AES128-GCM-SHA256,2048,128
Server Temp Key:        ECDH P-256 256 bits
TLS Server Name:        lobbywatch-staging.herokuapp.com

Document Path:          /de/daten/parlamentarier/364/Marianne%20Binder-Keller
Document Length:        Variable

Concurrency Level:      1
Time taken for tests:   28.359 seconds
Complete requests:      100
Failed requests:        0
Total transferred:      13607600 bytes
HTML transferred:       13578700 bytes
Requests per second:    3.53 [#/sec] (mean)
Time per request:       283.594 [ms] (mean)
Time per request:       283.594 [ms] (mean, across all concurrent requests)
Transfer rate:          468.58 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:      107  122   8.8    123     151
Processing:   144  162  11.3    164     183
Waiting:       39   46   7.8     45      90
Total:        252  284  19.6    287     319

Percentage of the requests served within a certain time (ms)
  50%    287
  66%    298
  75%    301
  80%    302
  90%    308
  95%    311
  98%    318
  99%    319
 100%    319 (longest request)
```

Will now adapt to all pages and polish.