@tpreusse 2021-04-25
Did a final performance comparison between the old code and the new code on heroku and vercel. Here the mean ms obtain by running `ab -l -n 50 -c 5` with various paths:

| path                                              | old heroku | new heroku | new vercel |
|---------------------------------------------------|-----------:|-----------:|-----------:|
| /de                                               |    994.743 |    310.837 |    210.955 |
| /de/artikel/vision-lobbyarbeit                    |    732.526 |    347.918 |    338.623 |
| /de/daten/parlamentarier                          |   2810.841 |    498.118 |    414.251 |
| /de/daten/parlamentarier/373/Martina%20Bircher    |    876.014 |    357.452 |    327.387 |
| /de/daten/zutrittsberechtigter                    |   2633.654 |   2945.593 |   1491.276 |
| /de/daten/zutrittsberechtigter/552/Vania%20Alleva |    811.100 |    905.742 |    760.619 |

The `/de/daten/zutrittsberechtigter` path, the related detail paths and 7 other pages are not yet migrated to static props. Generally we have 2x to 4x performance improvements for the optimised paths.

The remaining 9 pages will be easy to migrate but I'd prefer to launch as is and make sure static props work well in production too. So I will merge and deploy on Heroku tonight.

On the mean ms measured by `ab` Vercel isn't that big of an improvement. Also they currently do not (or no longer) have an edge cache in Zurich. But the cache in Paris and being able to locate the cloud functions there too is significantly faster for time to first byte (TTFB).  Even the the best case Heroku in Dublin remains around 50ms TTFB while Vercel get's down to around 30ms. So it's not a crazy huge benefit to migrate to Vercel but I still want to try it sometime next week.