@tpreusse 2021-04-25
One risk discovered: the new code can now prefetch code and data for pages that have static props. By default it does so quite aggressively—when ever a link enters the viewport. I've disabled that behaviour in lists and the footer. However hovering a link still prefetch it including the data and associated backend request. By hovering over all parliamentarians you can trigger 246 requests 🙈 

At the same time—when hovering just the one you want—navigation becomes really fast with hover prefetch. Anyhow curious to see it live in action while monitoring the app for timeouts.